Moved interfaces defined in types/auth.ts to entities/auth.ts. Updated imports that depended on types/auth.ts.
