+++ entities/auth.ts
@@ -1,14 +1,30 @@
-export interface AuthResponse {
-    token: string;
-    user: User;
-}
-
 export interface User {
-    id: number;
-    name: string;
+    id: string;
     email: string;
+    name: string;
+    role: 'user' | 'admin';
+}
+  
+export interface AuthState {
+    user: User | null;
+    token: string | null;
+    isAuthenticated: boolean;
+    isLoading: boolean;
+    error: string | null;
 }
 
-export interface MockUser extends User {
-    password: string
+export interface LoginCredentials {
+    email: string;
+    password: string;
 }
+
+export interface RegisterCredentials extends LoginCredentials {
+    name: string;
+}
+
+export interface AuthContextType extends AuthState {
+    login: (credentials: LoginCredentials) => Promise<void>;
+    register: (credentials: RegisterCredentials) => Promise<void>;
+    logout: () => void;
+    clearError: () => void;
+}
